refactor(api): move admin disable route to src/team
