+++ api/tests/unit/application/memberships/index_test.js
@@ -1,54 +1,9 @@
 import * as moduleUnderTest from '../../../../lib/application/memberships/index.js';
-import { membershipController } from '../../../../lib/application/memberships/membership-controller.js';
 import { securityPreHandlers } from '../../../../src/shared/application/security-pre-handlers.js';
+import { membershipController } from '../../../../src/team/application/membership/membership.controller.js';
 import { expect, HttpTestServer, sinon } from '../../../test-helper.js';
 
 describe('Unit | Router | membership-router', function () {
-  describe('POST /api/admin/memberships/{id}/disable', function () {
-    it('should return 204 if user is Pix Admin', async function () {
-      // given
-      sinon.stub(securityPreHandlers, 'hasAtLeastOneAccessOf').returns(() => true);
-      sinon.stub(membershipController, 'disable').callsFake((request, h) => h.response().code(204));
-
-      const httpTestServer = new HttpTestServer();
-      await httpTestServer.register(moduleUnderTest);
-      const membershipId = 123;
-
-      // when
-      const response = await httpTestServer.request('POST', `/api/admin/memberships/${membershipId}/disable`);
-
-      // then
-      expect(response.statusCode).to.equal(204);
-      expect(membershipController.disable).to.have.been.called;
-    });
-
-    it('returns forbidden access if admin member has CERTIF role', async function () {
-      // given
-      sinon.stub(membershipController, 'disable');
-
-      sinon.stub(securityPreHandlers, 'checkAdminMemberHasRoleCertif').callsFake((request, h) => h.response(true));
-      sinon
-        .stub(securityPreHandlers, 'checkAdminMemberHasRoleSuperAdmin')
-        .callsFake((request, h) => h.response({ errors: new Error('forbidden') }).code(403));
-      sinon
-        .stub(securityPreHandlers, 'checkAdminMemberHasRoleSupport')
-        .callsFake((request, h) => h.response({ errors: new Error('forbidden') }).code(403));
-      sinon
-        .stub(securityPreHandlers, 'checkAdminMemberHasRoleMetier')
-        .callsFake((request, h) => h.response({ errors: new Error('forbidden') }).code(403));
-
-      const httpTestServer = new HttpTestServer();
-      await httpTestServer.register(moduleUnderTest);
-
-      // when
-      const response = await httpTestServer.request('POST', '/api/admin/memberships/1/disable');
-
-      // then
-      expect(response.statusCode).to.equal(403);
-      expect(membershipController.disable).to.have.not.been.called;
-    });
-  });
-
   describe('POST /api/memberships/{id}/disable', function () {
     it('should return 204 if user is admin in organization', async function () {
       // given
